refactored average score to be float

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -6,7 +6,7 @@ CREATE PROCEDURE ComputeAverageScoreForUser (
     IN user_id INT
 )
 BEGIN
-    DECLARE avg_score DECIMAL(10,2);
+    DECLARE avg_score FLOAT;
 
     SELECT AVG(score) INTO avg_score FROM corrections WHERE user_id = user_id;
